feat: add repo view model and reduce warnings and make minimum deployment target  14.
